+++ utils/ktor-client/src/commonMain/kotlin/ApiInvoker.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2024-2025 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.utils.ktor
+
+import io.ktor.client.HttpClient
+
+
+interface ApiInvoker<Api> {
+    suspend operator fun <R> invoke(action: suspend Api.() -> R): R
+}
+
+fun <Api> ApiInvoker(
+    client: WrapperHttpClient,
+    getApi: (HttpClient) -> Api,
+): ApiInvoker<Api> = object : ApiInvoker<Api> {
+    override suspend fun <R> invoke(action: suspend Api.() -> R): R =
+        client.use {
+            action(
+                getApi(this),
+            )
+        }
+}
